+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Heretic skins
   * Void Reaver skins
   * Voidling skins
+  * Mini Mushrum skins
 
 * Changed:
   * Made Artificer cloth visible when using different model skin